refactor, model example related models to javagen
