+++ fluentgen/src/main/java/com/azure/autorest/fluent/mapper/ExampleParser.java
@@ -12,18 +12,18 @@ import com.azure.autorest.fluent.model.clientmodel.FluentResourceCollection;
 import com.azure.autorest.fluent.model.clientmodel.FluentResourceModel;
 import com.azure.autorest.fluent.model.clientmodel.FluentStatic;
 import com.azure.autorest.fluent.model.clientmodel.MethodParameter;
-import com.azure.autorest.fluent.model.clientmodel.ModelProperty;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.ClientModelNode;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.ExampleNode;
+import com.azure.autorest.model.clientmodel.ModelProperty;
+import com.azure.autorest.model.clientmodel.examplemodel.ClientModelNode;
+import com.azure.autorest.model.clientmodel.examplemodel.ExampleNode;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentClientMethodExample;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentCollectionMethodExample;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentMethodExample;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentResourceCreateExample;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentResourceUpdateExample;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.ListNode;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.LiteralNode;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.MapNode;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.ObjectNode;
+import com.azure.autorest.model.clientmodel.examplemodel.ListNode;
+import com.azure.autorest.model.clientmodel.examplemodel.LiteralNode;
+import com.azure.autorest.model.clientmodel.examplemodel.MapNode;
+import com.azure.autorest.model.clientmodel.examplemodel.ObjectNode;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.ParameterExample;
 import com.azure.autorest.fluent.model.clientmodel.fluentmodel.create.DefinitionStage;
 import com.azure.autorest.fluent.model.clientmodel.fluentmodel.create.DefinitionStageBlank;
