+++ fluentgen/src/main/java/com/azure/autorest/fluent/model/clientmodel/fluentmodel/FluentInterfaceStage.java
@@ -3,7 +3,7 @@
 
 package com.azure.autorest.fluent.model.clientmodel.fluentmodel;
 
-import com.azure.autorest.fluent.model.clientmodel.ModelProperty;
+import com.azure.autorest.model.clientmodel.ModelProperty;
 import com.azure.autorest.fluent.model.clientmodel.fluentmodel.method.FluentMethod;
 import com.azure.autorest.model.clientmodel.ClientMethodParameter;
