+++ fluentgen/src/main/java/com/azure/autorest/fluent/template/FluentExampleTemplate.java
@@ -6,18 +6,18 @@ package com.azure.autorest.fluent.template;
 import com.azure.autorest.extension.base.plugin.PluginLogger;
 import com.azure.autorest.fluent.FluentGen;
 import com.azure.autorest.fluent.model.clientmodel.FluentStatic;
-import com.azure.autorest.fluent.model.clientmodel.ModelProperty;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.ClientModelNode;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.ExampleNode;
+import com.azure.autorest.model.clientmodel.ModelProperty;
+import com.azure.autorest.model.clientmodel.examplemodel.ClientModelNode;
+import com.azure.autorest.model.clientmodel.examplemodel.ExampleNode;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentCollectionMethodExample;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentExample;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentMethodExample;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentResourceCreateExample;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.FluentResourceUpdateExample;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.ListNode;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.LiteralNode;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.MapNode;
-import com.azure.autorest.fluent.model.clientmodel.examplemodel.ObjectNode;
+import com.azure.autorest.model.clientmodel.examplemodel.ListNode;
+import com.azure.autorest.model.clientmodel.examplemodel.LiteralNode;
+import com.azure.autorest.model.clientmodel.examplemodel.MapNode;
+import com.azure.autorest.model.clientmodel.examplemodel.ObjectNode;
 import com.azure.autorest.fluent.model.clientmodel.examplemodel.ParameterExample;
 import com.azure.autorest.model.clientmodel.ClassType;
 import com.azure.autorest.model.clientmodel.ClientModel;
