+++ javagen/src/main/java/com/azure/autorest/model/clientmodel/ModelProperty.java
@@ -1,13 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.autorest.fluent.model.clientmodel;
+package com.azure.autorest.model.clientmodel;
 
-import com.azure.autorest.fluent.util.FluentUtils;
-import com.azure.autorest.model.clientmodel.ClientModelProperty;
-import com.azure.autorest.model.clientmodel.ClientModelPropertyAccess;
-import com.azure.autorest.model.clientmodel.ClientModelPropertyReference;
-import com.azure.autorest.model.clientmodel.IType;
+import com.azure.autorest.util.ModelUtil;
 
 import java.util.Collections;
 import java.util.List;
@@ -88,7 +84,7 @@ public class ModelProperty {
         if (property instanceof ClientModelProperty) {
             ClientModelProperty clientModelProperty = (ClientModelProperty) property;
             if (!clientModelProperty.getNeedsFlatten()) {
-                return FluentUtils.splitFlattenedSerializedName(clientModelProperty.getSerializedName());
+                return ModelUtil.splitFlattenedSerializedName(clientModelProperty.getSerializedName());
             } else {
                 return Collections.singletonList(clientModelProperty.getSerializedName());
             }
