+++ javagen/src/main/java/com/azure/autorest/model/clientmodel/examplemodel/ClientModelNode.java
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package com.azure.autorest.fluent.model.clientmodel.examplemodel;
+package com.azure.autorest.model.clientmodel.examplemodel;
 
-import com.azure.autorest.fluent.model.clientmodel.ModelProperty;
+import com.azure.autorest.model.clientmodel.ModelProperty;
 import com.azure.autorest.model.clientmodel.ClientModel;
 import com.azure.autorest.model.clientmodel.IType;
